Moved Sensor to the correct class
