+++ include/sensors/sensor.h
@@ -6,14 +6,19 @@
 
 class Sensor {
 public:
-  Sensor();
-  // const char* getType() const;
-  virtual const char* getValue() = 0;
-protected:
-  // Sensor(const char* type);
-  // char type[TYPE_BUFFER + 1];
-  // char value[VALUE_BUFFER + 1];
+  Sensor(const char* type) {
+    strncpy(this->type, type, 32); this->type[32] = 0;
+    memset(this->value, 0, 33);
+  }
+  const char* getType() const { return this->type; }
+  const char* getValue() const { return this->value; }
+  void setValue(const char* new_value) {
+    strncpy(this->value, new_value, 32);
+    this->value[32] = 0;
+  }
 private:
+  char type[33];
+  char value[33];
 };
 
 #endif
