+++ src/hub.cpp
@@ -6,6 +6,7 @@
 #include <ArduinoJson.h>
 #include "hub.h"
 #include "config.h"
+#include "sensors/sensor.h"
 
 Hub::Hub() {
   WiFi.mode(WIFI_AP_STA);
